Move WakeAction into app module
