+++ src/normal_mode.rs
@@ -16,12 +16,11 @@ use embedded_io_async::Read;
 use esp_hal::gpio::{Input, InputConfig, Output, Pull};
 use esp_println::println;
 
-use crate::app::AppContext;
+use crate::app::{AppContext, WakeAction};
 use crate::battery;
 use crate::button::wait_for_press;
 use crate::config::Config;
 use crate::error_image;
-use crate::hardware::WakeAction;
 use crate::panel::{Panel, PanelColor};
 use crate::rtc_persisted::RtcPersisted;
 use crate::sht40;
